Move Builder to its own module
